Add a guard to detect conflicting Flyway seqnos
Add a guard file with fixed header and a one-liner which is a
deterministic hash of all flyway scripts in the folder. The
guard file is checked every time test is run, and can be updated
with the :updateFlywayGuard task.

When concurrent schema changes are committed, only one of them will
succeed. Github will label all other PRs as out of date and request
reconciliation. This happens even if the sequence numbers do not
conflict.