Reassign xml namespace tags in Fee extensions
Nomulus supports multiple Fee extensions, and use hardcoded
version-specific namespace tags for each version. Previously the mapping
is as follows:
- Draft 06 --> 'fee'
- Draft 11 --> 'fee11'
- Draft 12 --> 'fee12'

When we added support for the final standard in RFC 8748, we assigned it
'fee_1_00'.

Some customers are having trouble with this new tag. We propose to
change the mappings to:

- Draft 06 --> 'fee06'
- Draft 11 --> 'fee11'
- Draft 12 --> 'fee12'
- Standard 1.0 -> 'fee'

Note that this only affects clients whose XML parsing is not
namespace-aware. Our XML is valid before AND after this change.